requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
filelock==3.13.4
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.3.1
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
regex==2024.4.16
Sastrawi==1.0.1
sympy==1.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
Werkzeug==3.0.2
